Undo api override, use component override in BookAppointmentDetails
