+++ src/pages/Appointments/BookAppointment/BookAppointmentDetails.tsx
@@ -1,3 +1,4 @@
+import { useMutation } from "@tanstack/react-query";
 import { useAtom } from "jotai";
 import { ArrowLeft, ArrowRight } from "lucide-react";
 import { navigate } from "raviger";
@@ -9,6 +10,7 @@ import { scheduleServiceTypeAtom } from "@/atoms/scheduleServiceTypeAtom";
 import { Button } from "@/components/ui/button";
 import { Drawer, DrawerContent, DrawerTrigger } from "@/components/ui/drawer";
 
+import { register } from "@/lib/override";
 import { AppointmentSlotPicker } from "@/pages/Appointments/BookAppointment/AppointmentSlotPicker";
 import useCurrentFacility from "@/pages/Facility/utils/useCurrentFacility";
 import { TagConfig } from "@/types/emr/tagConfig/tagConfig";
@@ -16,18 +18,19 @@ import scheduleApi from "@/types/scheduling/scheduleApi";
 import mutate from "@/Utils/request/mutate";
 
 import { ScheduleResourceFormState } from "@/components/Schedule/ResourceSelector";
-import { useApiMutation } from "@/hooks/useApiMutation";
 import { Appointment } from "@/types/scheduling/schedule";
 import { AppointmentDateSelection } from "./AppointmentDateSelection";
 import { AppointmentFormSection } from "./AppointmentFormSection";
 
-export const BookAppointmentDetails = ({
-  patientId,
-  onSuccess,
-}: {
+export interface BookAppointmentDetailsProps {
   patientId: string;
   onSuccess?: () => void;
-}) => {
+}
+
+const BookAppointmentDetailsBase = ({
+  patientId,
+  onSuccess,
+}: BookAppointmentDetailsProps) => {
   const { t } = useTranslation();
 
   const { facilityId } = useCurrentFacility();
@@ -70,8 +73,8 @@ export const BookAppointmentDetails = ({
     }
   };
 
-  const { mutateAsync: createAppointment, isPending: isCreating } =
-    useApiMutation({
+  const { mutateAsync: createAppointment, isPending: isCreating } = useMutation(
+    {
       mutationFn: mutate(scheduleApi.slots.createAppointment, {
         pathParams: { facilityId, slotId: selectedSlotId ?? "" },
       }),
@@ -82,7 +85,8 @@ export const BookAppointmentDetails = ({
           `/facility/${facilityId}/patient/${patientId}/appointments/${data.id}?showSuccess=true`,
         );
       },
-    });
+    },
+  );
 
   const handleSubmit = async () => {
     if (!selectedResource || !selectedSlotId) {
@@ -238,3 +242,8 @@ export const BookAppointmentDetails = ({
     </div>
   );
 };
+
+export const BookAppointmentDetails = register(
+  "BookAppointmentDetails",
+  BookAppointmentDetailsBase,
+);
